+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # Change Log
 
+### v1.1 (06.09.2023)
+ - Adding Manifest to give more information about the module
+ - Importing depenency module Microsoft.Dynamics.Nav.Model.Tools directly
+ - Change LanguageId from String to Integer (as of Windows LCID) for better clarity and extensibility
+ - Removing LanguageOption.psm1 file as no need to store language options any more
+ - DeepL API usage implementation
+ - DeepL API key storage implementation
+ - Updated Help and README
+ - Rename functions to use approved verbs
+ - Code refactoring and optimizations
+
 ### v1.0 (30.08.2023)
-Core module created with GitHub repository
+ - Core module created with GitHub repository